feat(info.ts): add localization for bot information command (#2b5443cc)
In the 'info.ts' file, the command description for the bot's information has been updated to use localization. This has been done for both English and French languages. The corresponding translation strings have been added to the 'en.json' and 'fr.json' files under the 'info.cmds' key. This change allows the bot's information command to be displayed in the user's selected language.